FIX: Hidden layer dimension assignment error - Corrected!
In the previous commit, I claimed to have fixed this. It turns out that I was lying. The input dimension of the decoder layer is meant to be the same dimension as the output space, and should not be the size of the hidden layer. I had coded the system to work in the wrong way, which caused dimensionality mismatches when trying to assign hidden layer dimension sizes not equal to the size of the input dimension.
This commit contains the corrections to that error. This *should* make things work just peachy on the GPU.